Refactor EuclideanGraph to abstract Graph code
I refactored EuclideanGraph so that base Graph code is no longer tightly coupled with the Euclidean graph code. This will allow the graph code to be extended for use in either a tree, or other implementations
