+++ src/EuclideanGraph.java
@@ -13,31 +13,18 @@ import java.util.Random;
 import java.util.ArrayList;
 import java.util.HashSet;
 
-public abstract class EuclideanGraph extends Visualizer {
-    private String title;
-    protected ArrayList<Vertex> vertices;
-    protected HashSet<Edge> edges;
-    private Color titleColor, backgroundColor;
-    protected Color vertexColor, selectedVertexColor;
-    protected Color edgeColor, selectedEdgeColor;
-    private final static int MAX_X = Visualizer.DRAW_WIDTH;
-    // subtract 100 for drawing of title
-    private final static int MAX_Y = Visualizer.DRAW_HEIGHT - 100;
-
+public abstract class EuclideanGraph extends Graph {
+    protected Color edgeColor, vertexColor;
 
     public EuclideanGraph(String title) {
-        this.title = title;
+        super(title);
+        buildGraph();
+    }
 
-        // Set default colors
-        backgroundColor = Color.BLACK;
-        titleColor = Color.WHITE;
-        vertexColor = Color.CYAN;
-        selectedVertexColor = Color.GREEN;
-        edgeColor = Color.MAGENTA;
-        selectedEdgeColor = Color.ORANGE;
-
-        vertices = new ArrayList<Vertex>();
-        edges = new HashSet<Edge>();
+    public EuclideanGraph(String title, Color backgroundColor, Color titleColor, Color vertexColor, Color edgeColor) {
+        super(title, backgroundColor, titleColor);
+        this.vertexColor = vertexColor;
+        this.edgeColor = edgeColor;
         buildGraph();
     }
 
@@ -95,9 +82,8 @@ public abstract class EuclideanGraph extends Visualizer {
             // This optimizes number of nodes
             if (!edges.contains(e) && e.getWeight() < MAX_Y / 5 && !vertexA.equals(vertexB)) {
                 edges.add(e);
-                // Todo: consider making Edge() {constructor} call Vertex.addEdge
-                vertexA.addEdge(e);
-                vertexB.addEdge(e);
+                vertexA.addEdgeWeightSorted(e);
+                vertexB.addEdgeWeightSorted(e);
                 if (verticeSet.contains(vertexA)) {
                     verticeSet.remove(vertexA);
                 }
@@ -117,190 +103,5 @@ public abstract class EuclideanGraph extends Visualizer {
             vertices.remove(v);
         }
         
-    }
-
-    @Override
-    public void paintVisualization(Graphics2D g2d) {
-        g2d.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-        // Draw Background
-        g2d.setColor(backgroundColor);
-        g2d.fillRect(0, 0, Visualizer.DRAW_WIDTH, Visualizer.DRAW_HEIGHT);
-
-        // Draw font
-        Font font = new Font("Helvetica", Font.BOLD, 60);
-        g2d.setFont(font);
-        g2d.setColor(titleColor);
-        FontMetrics metr = this.getFontMetrics(font);
-        g2d.drawString(title, (Visualizer.DRAW_WIDTH - metr.stringWidth(title)) / 2, 100);
-
-        // Paint vertices
-        for (Edge e : edges) {
-            e.paint(g2d);
-        }
-
-        // Paint vertices second to overlap color of edges
-        for (Vertex v : vertices) {
-            v.paint(g2d);
-        }
-
-        
-    }
-
-
-    protected class Vertex {
-        private ArrayList<Edge> edges;
-        private int x, y;
-        private Color color;
-        private int diameter;
-
-        public Vertex(int x, int y) {
-            this.x = x;
-            this.y = y;
-            diameter = 10;
-            edges = new ArrayList<Edge>();
-        }
-
-        public Vertex(int x, int y, Color color) {
-            this(x, y);
-            this.color = color;
-        }
-
-        public int getX() {
-            return x;
-        }
-
-        public int getY() {
-            return y;
-        }
-
-        public void addEdge(Edge e) {
-            // Add edges in sorted order by weight
-            if (edges.isEmpty()) {
-                // There is nothing in the list, so add the edge in the first spot
-                edges.add(e);
-            } else {
-                for (int i = 0; i < edges.size(); i++) {
-                    if (e.getWeight() < edges.get(i).getWeight()) {
-                        edges.add(i, e);
-                        return;
-                    }
-                }
-                // This edge is the longer than the rest of the list
-                edges.add(e);
-            }
-        }
-
-        public ArrayList<Edge> getEdges() {
-            return edges;
-        }
-
-        public Color getColor() {
-            return color;
-        }
-
-        public void setColor(Color color) {
-            this.color = color;
-        }
-
-        public int getDiameter() {
-            return diameter;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (obj == null || !(obj instanceof Vertex)) return false;
-            Vertex v = (Vertex) obj;
-            return (x == v.getX() && y ==v.getY());
-        }
-
-        public void paint(Graphics2D g2d) {
-            g2d.setColor(color);
-            g2d.fillOval(x, y, diameter, diameter);
-        }
-
-    }
-
-    protected class Edge {
-        double weight;
-        Vertex a, b;
-        Color color;
-
-        public Edge(Vertex a, Vertex b) {
-            this.a = a;
-            this.b = b;
-            weight = Math.sqrt(Math.pow(a.getX() - b.getX(), 2) + Math.pow(a.getY() - b.getY(), 2));
-        }
-
-        public Edge(Vertex a, Vertex b, Color color) {
-            this(a, b);
-            this.color = color;
-        }
-
-        public Edge(Vertex a, Vertex b, double weight) {
-            this.a = a;
-            this.b = b;
-            this.weight = weight;
-        }
-
-        public Edge(Vertex a, Vertex b, double weight, Color color) {
-            this(a, b, weight);
-            this.color = color;
-        }
-
-        public double getWeight() {
-            return weight;
-        }
-
-        public Vertex getVertexA() {
-            return a;
-        }
-
-        public Vertex getVertexB() {
-            return  b;
-        }
-
-        public Color getColor() {
-            return color;
-        }
-
-        public void setColor(Color color) {
-            this.color = color;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (obj == null || !(obj instanceof Edge)) return false;
-            Edge e = (Edge) obj;
-            return ((a.equals(e.getVertexA()) || a.equals(e.getVertexB()) ) && b.equals(e.getVertexA()) || b.equals(e.getVertexB()));
-        }
-
-        public void paint(Graphics2D g2d) {
-            // Save the old stroke so we can reset it
-            Stroke old = g2d.getStroke();
-            g2d.setStroke(new BasicStroke(3));
-            g2d.setColor(color);
-            int offset = a.getDiameter() /2;
-            g2d.drawLine(a.getX() + offset, a.getY() + offset, b.getX() + offset, b.getY() + offset);
-
-            // Prep Font for drawing weight
-            Font font = new Font("Helvetica", Font.PLAIN, 15);
-            g2d.setFont(font);
-            g2d.setColor(titleColor);
-            // Build weight as a string, with a max length of 5 characters
-            String weightStr = Double.toString(weight).subSequence(0, Math.min(Double.toString(weight).length(), 5)).toString();
-            //g2d.drawString(weightStr, (a.getX() + b.getX()) / 2, (a.getY() + b.getY()) / 2);
-            g2d.setStroke(old);
-        }
-
-        public boolean intersects(Edge e) {
-            if (Math.max(a.getX(), b.getX()) > Math.min(e.getVertexA().getX(), e.getVertexB().getX()) 
-                && Math.min(a.getX(), b.getX()) < Math.max(e.getVertexA().getX(), e.getVertexB().getX())) {
-                if (Math.max(a.getY(), b.getY()) > Math.min(e.getVertexA().getY(), e.getVertexB().getY()) 
-                    && Math.min(a.getY(), b.getY()) < Math.max(e.getVertexA().getY(), e.getVertexB().getY())) {
-                    return true;
-                }
-            }
-            return false;
-        }
     }
 }
